Module Yocaml_irmin

Allows you to define a Runtime that uses an Irmin Store (e.g. Git) as a compilation target.

API

val execute : (module Yocaml_irmin__Runtime.RUNTIME) -> (module Irmin.S with type Schema.Branch.t = string and type Schema.Contents.t = string and type Schema.Path.t = string list) -> ?branch:string -> ?author:string -> ?author_email:string -> Irmin.config -> 'a Yocaml.Effect.t -> 'a Lwt.t

Executes a YOCaml program using a given Runtime for processing with Source and using an Irmin Store as compilation target.