Module Yocaml_irmin

Allows you to define a Runtime that uses an Irmin Store (e.g. Git) as a compilation target.

API

val execute : (module Yocaml.Runtime.RUNTIME) -> (module Irmin.S with type branch = string and type contents = string and type key = string list) -> (module Yocaml_irmin__.Runtime.LWT_RUN) -> ?⁠branch:string -> ?⁠author:string -> ?⁠author_email:string -> Irmin.config -> 'a Yocaml.Effect.t -> 'a

Executes a YOCaml program using a given Runtime for processing with Source and using an Irmin Store as compilation target.